Refactor: Centralize mock DOM utilities in testHelpers.js
I moved core DOM mocking utilities (createMockElement, createMockDocument, createMockWindow) from flint.test.js to client/testHelpers.js.

I then updated flint.test.js and debug.test.js to use these shared utilities, reducing code duplication and improving consistency in your test setup.

This is a first step towards standardizing mocking strategies across your client-side tests. All tests pass with these changes.
